pyproject.toml
[project]
name = "budget-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.4",
    "motor>=3.6.0",
    "pydantic-settings>=2.6.1",
]

[tool.uv]
dev-dependencies = ["black>=24.10.0", "mypy>=1.12.0", "ruff>=0.7.0"]

[tool.black]
line-length = 88

[tool.ruff]
lint.select = ["E", "F", "I"]
lint.fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
plugins = ["pydantic.mypy"]

disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
